# Optimize Slice Allocations ( ~ -19% byte/op, -14% ns/op, ~ -11% allocs )