--- a/internal/store/configmap.go
+++ b/internal/store/configmap.go
@@ -142,8 +142,14 @@ func wrapConfigMapFunc(f func(*v1.ConfigMap) *metric.Family) func(interface{}) *
 		metricFamily := f(configMap)
 
 		for _, m := range metricFamily.Metrics {
-			m.LabelKeys = append(descConfigMapLabelsDefaultLabels, m.LabelKeys...)
-			m.LabelValues = append([]string{configMap.Namespace, configMap.Name}, m.LabelValues...)
+			commonLabelKeys := make([]string, 0, len(descConfigMapLabelsDefaultLabels)+len(m.LabelKeys))
+			commonLabelValues := make([]string, 0, len(descConfigMapLabelsDefaultLabels)+len(m.LabelValues))
+
+			commonLabelKeys = append(commonLabelKeys, descConfigMapLabelsDefaultLabels...)
+			commonLabelValues = append(commonLabelValues, configMap.Namespace, configMap.Name)
+
+			m.LabelKeys = append(commonLabelKeys, m.LabelKeys...)
+			m.LabelValues = append(commonLabelValues, m.LabelValues...)
 		}
 
 		return metricFamily