--- a/internal/store/daemonset.go
+++ b/internal/store/daemonset.go
@@ -258,8 +258,14 @@ func wrapDaemonSetFunc(f func(*v1.DaemonSet) *metric.Family) func(interface{}) *
 		metricFamily := f(daemonSet)
 
 		for _, m := range metricFamily.Metrics {
-			m.LabelKeys = append(descDaemonSetLabelsDefaultLabels, m.LabelKeys...)
-			m.LabelValues = append([]string{daemonSet.Namespace, daemonSet.Name}, m.LabelValues...)
+			commonLabelKeys := make([]string, 0, len(descDaemonSetLabelsDefaultLabels)+len(m.LabelKeys))
+			commonLabelValues := make([]string, 0, len(descDaemonSetLabelsDefaultLabels)+len(m.LabelValues))
+
+			commonLabelKeys = append(commonLabelKeys, descDaemonSetLabelsDefaultLabels...)
+			commonLabelValues = append(commonLabelValues, daemonSet.Namespace, daemonSet.Name)
+
+			m.LabelKeys = append(commonLabelKeys, m.LabelKeys...)
+			m.LabelValues = append(commonLabelValues, m.LabelValues...)
 		}
 
 		return metricFamily