--- a/internal/store/deployment.go
+++ b/internal/store/deployment.go
@@ -314,8 +314,14 @@ func wrapDeploymentFunc(f func(*v1.Deployment) *metric.Family) func(interface{})
 		metricFamily := f(deployment)
 
 		for _, m := range metricFamily.Metrics {
-			m.LabelKeys = append(descDeploymentLabelsDefaultLabels, m.LabelKeys...)
-			m.LabelValues = append([]string{deployment.Namespace, deployment.Name}, m.LabelValues...)
+			commonLabelKeys := make([]string, 0, len(descDeploymentLabelsDefaultLabels)+len(m.LabelKeys))
+			commonLabelValues := make([]string, 0, len(descDeploymentLabelsDefaultLabels)+len(m.LabelValues))
+
+			commonLabelKeys = append(commonLabelKeys, descDeploymentLabelsDefaultLabels...)
+			commonLabelValues = append(commonLabelValues, deployment.Namespace, deployment.Name)
+
+			m.LabelKeys = append(commonLabelKeys, m.LabelKeys...)
+			m.LabelValues = append(commonLabelValues, m.LabelValues...)
 		}
 
 		return metricFamily