--- a/internal/store/endpoint.go
+++ b/internal/store/endpoint.go
@@ -182,8 +182,14 @@ func wrapEndpointFunc(f func(*v1.Endpoints) *metric.Family) func(interface{}) *m
 		metricFamily := f(endpoint)
 
 		for _, m := range metricFamily.Metrics {
-			m.LabelKeys = append(descEndpointLabelsDefaultLabels, m.LabelKeys...)
-			m.LabelValues = append([]string{endpoint.Namespace, endpoint.Name}, m.LabelValues...)
+			commonLabelKeys := make([]string, 0, len(descEndpointLabelsDefaultLabels)+len(m.LabelKeys))
+			commonLabelValues := make([]string, 0, len(descEndpointLabelsDefaultLabels)+len(m.LabelValues))
+
+			commonLabelKeys = append(commonLabelKeys, descEndpointLabelsDefaultLabels...)
+			commonLabelValues = append(commonLabelValues, endpoint.Namespace, endpoint.Name)
+
+			m.LabelKeys = append(commonLabelKeys, m.LabelKeys...)
+			m.LabelValues = append(commonLabelValues, m.LabelValues...)
 		}
 
 		return metricFamily