--- a/internal/store/job.go
+++ b/internal/store/job.go
@@ -387,8 +387,14 @@ func wrapJobFunc(f func(*v1batch.Job) *metric.Family) func(interface{}) *metric.
 		metricFamily := f(job)
 
 		for _, m := range metricFamily.Metrics {
-			m.LabelKeys = append(descJobLabelsDefaultLabels, m.LabelKeys...)
-			m.LabelValues = append([]string{job.Namespace, job.Name}, m.LabelValues...)
+			commonLabelKeys := make([]string, 0, len(descJobLabelsDefaultLabels)+len(m.LabelKeys))
+			commonLabelValues := make([]string, 0, len(descJobLabelsDefaultLabels)+len(m.LabelValues))
+
+			commonLabelKeys = append(commonLabelKeys, descJobLabelsDefaultLabels...)
+			commonLabelValues = append(commonLabelValues, job.Namespace, job.Name)
+
+			m.LabelKeys = append(commonLabelKeys, m.LabelKeys...)
+			m.LabelValues = append(commonLabelValues, m.LabelValues...)
 		}
 
 		return metricFamily