--- a/internal/store/lease.go
+++ b/internal/store/lease.go
@@ -97,8 +97,14 @@ func wrapLeaseFunc(f func(*coordinationv1.Lease) *metric.Family) func(interface{
 		metricFamily := f(lease)
 
 		for _, m := range metricFamily.Metrics {
-			m.LabelKeys = append(descLeaseLabelsDefaultLabels, m.LabelKeys...)
-			m.LabelValues = append([]string{lease.Name}, m.LabelValues...)
+			commonLabelKeys := make([]string, 0, len(descLeaseLabelsDefaultLabels)+len(m.LabelKeys))
+			commonLabelValues := make([]string, 0, len(descLeaseLabelsDefaultLabels)+len(m.LabelValues))
+
+			commonLabelKeys = append(commonLabelKeys, descLeaseLabelsDefaultLabels...)
+			commonLabelValues = append(commonLabelValues, lease.Name)
+
+			m.LabelKeys = append(commonLabelKeys, m.LabelKeys...)
+			m.LabelValues = append(commonLabelValues, m.LabelValues...)
 		}
 
 		return metricFamily