--- a/internal/store/namespace.go
+++ b/internal/store/namespace.go
@@ -155,8 +155,14 @@ func wrapNamespaceFunc(f func(*v1.Namespace) *metric.Family) func(interface{}) *
 		metricFamily := f(namespace)
 
 		for _, m := range metricFamily.Metrics {
-			m.LabelKeys = append(descNamespaceLabelsDefaultLabels, m.LabelKeys...)
-			m.LabelValues = append([]string{namespace.Name}, m.LabelValues...)
+			commonLabelKeys := make([]string, 0, len(descNamespaceLabelsDefaultLabels)+len(m.LabelKeys))
+			commonLabelValues := make([]string, 0, len(descNamespaceLabelsDefaultLabels)+len(m.LabelValues))
+
+			commonLabelKeys = append(commonLabelKeys, descNamespaceLabelsDefaultLabels...)
+			commonLabelValues = append(commonLabelValues, namespace.Name)
+
+			m.LabelKeys = append(commonLabelKeys, m.LabelKeys...)
+			m.LabelValues = append(commonLabelValues, m.LabelValues...)
 		}
 
 		return metricFamily