--- a/internal/store/volumeattachment.go
+++ b/internal/store/volumeattachment.go
@@ -153,8 +153,14 @@ func wrapVolumeAttachmentFunc(f func(*storagev1.VolumeAttachment) *metric.Family
 		metricFamily := f(va)
 
 		for _, m := range metricFamily.Metrics {
-			m.LabelKeys = append(descVolumeAttachmentLabelsDefaultLabels, m.LabelKeys...)
-			m.LabelValues = append([]string{va.Name}, m.LabelValues...)
+			commonLabelKeys := make([]string, 0, len(descVolumeAttachmentLabelsDefaultLabels)+len(m.LabelKeys))
+			commonLabelValues := make([]string, 0, len(descVolumeAttachmentLabelsDefaultLabels)+len(m.LabelValues))
+
+			commonLabelKeys = append(commonLabelKeys, descVolumeAttachmentLabelsDefaultLabels...)
+			commonLabelValues = append(commonLabelValues, va.Name)
+
+			m.LabelKeys = append(commonLabelKeys, m.LabelKeys...)
+			m.LabelValues = append(commonLabelValues, m.LabelValues...)
 		}
 
 		return metricFamily