<!--  Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, read our contributor guidelines https://git.k8s.io/community/contributors/guide/pull-requests.md#the-pull-request-submit-process and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
3. Follow the instructions for writing a release note: https://git.k8s.io/community/contributors/guide/pull-requests.md#write-release-notes-if-needed
4. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

# **What this PR does / why we need it**:

- This PR Optimizes Slices Allocations in the `wrap<resource>Func()` by pre-allocate the required capacity and avoiding re-allocations.
- The slices are small, however, **two new slices** are allocated, and expanded, **for each metric family, of each resource**. 
- Pre-allocating slices capacity is possible because we know how many element we'll need for each slice. 

**How does this change affect the cardinality of KSM**: *(increases, decreases or does not change cardinality)*
- Does not affect cardinality.

#  Benchmarks:

Using ` make test-benchmark-compare`

### Compared to Master
<details>
<summary> Logs </summary>

```
./tests/compare_benchmarks.sh master

### Testing alloc-optimize
?   	k8s.io/kube-state-metrics/v2	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/internal/store
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkPodStore-16    	  106828	     11251 ns/op	   12816 B/op	     232 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/internal/store	2.354s
?   	k8s.io/kube-state-metrics/v2/pkg/allow	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/allowdenylist	0.411s
I0205 18:59:19.881358   94914 builder.go:232] Active resources: certificatesigningrequests,configmaps,cronjobs,daemonsets,deployments,endpoints,horizontalpodautoscalers,ingresses,jobs,leases,limitranges,mutatingwebhookconfigurations,namespaces,networkpolicies,nodes,persistentvolumeclaims,persistentvolumes,poddisruptionbudgets,pods,replicasets,replicationcontrollers,resourcequotas,secrets,services,statefulsets,storageclasses,validatingwebhookconfigurations,volumeattachments
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/app
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkKubeStateMetrics/GenerateMetrics-16         	       1	1004373526 ns/op	63972984 B/op	  602716 allocs/op
BenchmarkKubeStateMetrics/MakeRequests-16            	       1	3626366009 ns/op	1712.31 MB/s	14257183408 B/op	  509450 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/app	8.037s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/builder	0.450s
?   	k8s.io/kube-state-metrics/v2/pkg/builder/types	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/constant	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/customresource	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/metric
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkMetricWrite/value-1-16         	 2916366	       407.8 ns/op	     536 B/op	       7 allocs/op
BenchmarkMetricWrite/value-35.7-16      	 2276084	       528.0 ns/op	     536 B/op	       7 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metric	3.655s
?   	k8s.io/kube-state-metrics/v2/pkg/metric_generator	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metrics_store	0.477s
?   	k8s.io/kube-state-metrics/v2/pkg/metricshandler	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/optin	0.494s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/options	0.478s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/sharding	0.449s
?   	k8s.io/kube-state-metrics/v2/pkg/util/proc	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/version	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/watch	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/e2e	0.388s
?   	k8s.io/kube-state-metrics/v2/tests/e2e/framework	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/lib	0.507s

### Done testing alloc-optimize

### Testing master
Switched to branch 'master'
Your branch is ahead of 'origin/master' by 2 commits.
  (use "git push" to publish your local commits)
?   	k8s.io/kube-state-metrics/v2	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/internal/store
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkPodStore-16    	   90944	     13116 ns/op	   15888 B/op	     262 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/internal/store	1.852s
?   	k8s.io/kube-state-metrics/v2/pkg/allow	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/allowdenylist	0.340s
I0205 18:59:38.861400   95006 builder.go:240] Active resources: certificatesigningrequests,configmaps,cronjobs,daemonsets,deployments,endpoints,horizontalpodautoscalers,ingresses,jobs,leases,limitranges,mutatingwebhookconfigurations,namespaces,networkpolicies,nodes,persistentvolumeclaims,persistentvolumes,poddisruptionbudgets,pods,replicasets,replicationcontrollers,resourcequotas,secrets,services,statefulsets,storageclasses,validatingwebhookconfigurations,volumeattachments
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/app
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkKubeStateMetrics/GenerateMetrics-16         	       1	1001920205 ns/op	66873336 B/op	  636615 allocs/op
BenchmarkKubeStateMetrics/MakeRequests-16            	       1	3657135938 ns/op	1697.90 MB/s	14257066328 B/op	  509427 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/app	8.074s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/builder	0.553s
?   	k8s.io/kube-state-metrics/v2/pkg/builder/types	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/constant	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/customresource	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/metric
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkMetricWrite/value-1-16         	 2950290	       406.6 ns/op	     536 B/op	       7 allocs/op
BenchmarkMetricWrite/value-35.7-16      	 2294436	       527.8 ns/op	     536 B/op	       7 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metric	3.631s
?   	k8s.io/kube-state-metrics/v2/pkg/metric_generator	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metrics_store	0.309s
?   	k8s.io/kube-state-metrics/v2/pkg/metricshandler	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/optin	0.242s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/options	0.410s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/sharding	0.353s
?   	k8s.io/kube-state-metrics/v2/pkg/util/proc	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/version	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/watch	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/e2e	0.334s
?   	k8s.io/kube-state-metrics/v2/tests/e2e/framework	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/lib	0.461s

### Done testing master
Switched to branch 'alloc-optimize'
Your branch is up to date with 'fork/alloc-optimize'.

### Result
old=master new=alloc-optimize
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                        old ns/op      new ns/op      delta
BenchmarkPodStore-16                             13116          11251          -14.22%
BenchmarkKubeStateMetrics/GenerateMetrics-16     1001920205     1004373526     +0.24%
BenchmarkKubeStateMetrics/MakeRequests-16        3657135938     3626366009     -0.84%
BenchmarkMetricWrite/value-1-16                  407            408            +0.30%
BenchmarkMetricWrite/value-35.7-16               528            528            +0.04%

benchmark                                     old MB/s     new MB/s     speedup
BenchmarkKubeStateMetrics/MakeRequests-16     1697.90      1712.31      1.01x

benchmark                                        old allocs     new allocs     delta
BenchmarkPodStore-16                             262            232            -11.45%
BenchmarkKubeStateMetrics/GenerateMetrics-16     636615         602716         -5.32%
BenchmarkKubeStateMetrics/MakeRequests-16        509427         509450         +0.00%
BenchmarkMetricWrite/value-1-16                  7              7              +0.00%
BenchmarkMetricWrite/value-35.7-16               7              7              +0.00%

benchmark                                        old bytes       new bytes       delta
BenchmarkPodStore-16                             15888           12816           -19.34%
BenchmarkKubeStateMetrics/GenerateMetrics-16     66873336        63972984        -4.34%
BenchmarkKubeStateMetrics/MakeRequests-16        14257066328     14257183408     +0.00%
BenchmarkMetricWrite/value-1-16                  536             536             +0.00%
BenchmarkMetricWrite/value-35.7-16               536             536             +0.00%
```
</details>

#### results
```
### Result
old=master new=alloc-optimize
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                        old ns/op      new ns/op      delta
BenchmarkPodStore-16                             13116          11251          -14.22%
BenchmarkKubeStateMetrics/GenerateMetrics-16     1001920205     1004373526     +0.24%
BenchmarkKubeStateMetrics/MakeRequests-16        3657135938     3626366009     -0.84%
BenchmarkMetricWrite/value-1-16                  407            408            +0.30%
BenchmarkMetricWrite/value-35.7-16               528            528            +0.04%

benchmark                                     old MB/s     new MB/s     speedup
BenchmarkKubeStateMetrics/MakeRequests-16     1697.90      1712.31      1.01x

benchmark                                        old allocs     new allocs     delta
BenchmarkPodStore-16                             262            232            -11.45%
BenchmarkKubeStateMetrics/GenerateMetrics-16     636615         602716         -5.32%
BenchmarkKubeStateMetrics/MakeRequests-16        509427         509450         +0.00%
BenchmarkMetricWrite/value-1-16                  7              7              +0.00%
BenchmarkMetricWrite/value-35.7-16               7              7              +0.00%

benchmark                                        old bytes       new bytes       delta
BenchmarkPodStore-16                             15888           12816           -19.34%
BenchmarkKubeStateMetrics/GenerateMetrics-16     66873336        63972984        -4.34%
BenchmarkKubeStateMetrics/MakeRequests-16        14257066328     14257183408     +0.00%
BenchmarkMetricWrite/value-1-16                  536             536             +0.00%
BenchmarkMetricWrite/value-35.7-16               536             536             +0.00%
```



### Compared to Latest Release
<details>
<summary> Logs </summary>

```
### Testing alloc-optimize
?   	k8s.io/kube-state-metrics/v2	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/internal/store
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkPodStore-16    	  106470	     11287 ns/op	   12816 B/op	     232 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/internal/store	1.735s
?   	k8s.io/kube-state-metrics/v2/pkg/allow	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/allowdenylist	0.259s
I0205 18:59:57.085779   95106 builder.go:232] Active resources: certificatesigningrequests,configmaps,cronjobs,daemonsets,deployments,endpoints,horizontalpodautoscalers,ingresses,jobs,leases,limitranges,mutatingwebhookconfigurations,namespaces,networkpolicies,nodes,persistentvolumeclaims,persistentvolumes,poddisruptionbudgets,pods,replicasets,replicationcontrollers,resourcequotas,secrets,services,statefulsets,storageclasses,validatingwebhookconfigurations,volumeattachments
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/app
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkKubeStateMetrics/GenerateMetrics-16         	       1	1003332439 ns/op	63895760 B/op	  602677 allocs/op
BenchmarkKubeStateMetrics/MakeRequests-16            	       1	3730689662 ns/op	1664.43 MB/s	14256884576 B/op	  509375 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/app	8.054s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/builder	0.424s
?   	k8s.io/kube-state-metrics/v2/pkg/builder/types	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/constant	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/customresource	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/metric
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkMetricWrite/value-1-16         	 2967225	       404.1 ns/op	     536 B/op	       7 allocs/op
BenchmarkMetricWrite/value-35.7-16      	 2307228	       522.2 ns/op	     536 B/op	       7 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metric	3.652s
?   	k8s.io/kube-state-metrics/v2/pkg/metric_generator	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metrics_store	0.268s
?   	k8s.io/kube-state-metrics/v2/pkg/metricshandler	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/optin	0.202s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/options	0.259s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/sharding	0.337s
?   	k8s.io/kube-state-metrics/v2/pkg/util/proc	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/version	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/watch	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/e2e	0.320s
?   	k8s.io/kube-state-metrics/v2/tests/e2e/framework	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/lib	0.443s

### Done testing alloc-optimize

### Testing release-2.3
Switched to branch 'release-2.3'
Your branch is up to date with 'origin/release-2.3'.
I0205 19:00:13.391361   95193 builder.go:192] Active resources: certificatesigningrequests,configmaps,cronjobs,daemonsets,deployments,endpoints,horizontalpodautoscalers,ingresses,jobs,leases,limitranges,mutatingwebhookconfigurations,namespaces,networkpolicies,nodes,persistentvolumeclaims,persistentvolumes,poddisruptionbudgets,pods,replicasets,replicationcontrollers,resourcequotas,secrets,services,statefulsets,storageclasses,validatingwebhookconfigurations,volumeattachments
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkKubeStateMetrics/GenerateMetrics-16         	       1	1004029168 ns/op	66918864 B/op	  636653 allocs/op
BenchmarkKubeStateMetrics/MakeRequests-16            	       1	3674074324 ns/op	1690.08 MB/s	14257290016 B/op	  509309 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2	8.504s
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/internal/store
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkPodStore-16    	   91106	     13081 ns/op	   15888 B/op	     262 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/internal/store	1.860s
?   	k8s.io/kube-state-metrics/v2/pkg/allow	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/allowdenylist	0.324s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/builder	0.532s
?   	k8s.io/kube-state-metrics/v2/pkg/builder/types	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/constant	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/metric
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkMetricWrite/value-1-16         	 2925446	       408.5 ns/op	     536 B/op	       7 allocs/op
BenchmarkMetricWrite/value-35.7-16      	 2271388	       526.8 ns/op	     536 B/op	       7 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metric	3.620s
?   	k8s.io/kube-state-metrics/v2/pkg/metric_generator	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metrics_store	0.305s
?   	k8s.io/kube-state-metrics/v2/pkg/metricshandler	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/optin	0.294s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/options	0.356s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/sharding	0.376s
?   	k8s.io/kube-state-metrics/v2/pkg/util/proc	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/version	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/watch	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/e2e	0.301s
?   	k8s.io/kube-state-metrics/v2/tests/e2e/framework	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/lib	0.374s

### Done testing release-2.3
```
</details>

#### results
```
### Result
old=release-2.3 new=alloc-optimize
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                        old ns/op      new ns/op      delta
BenchmarkKubeStateMetrics/GenerateMetrics-16     1004029168     1003332439     -0.07%
BenchmarkKubeStateMetrics/MakeRequests-16        3674074324     3730689662     +1.54%
BenchmarkPodStore-16                             13081          11287          -13.71%
BenchmarkMetricWrite/value-1-16                  408            404            -1.08%
BenchmarkMetricWrite/value-35.7-16               527            522            -0.87%

benchmark                                     old MB/s     new MB/s     speedup
BenchmarkKubeStateMetrics/MakeRequests-16     1690.08      1664.43      0.98x

benchmark                                        old allocs     new allocs     delta
BenchmarkKubeStateMetrics/GenerateMetrics-16     636653         602677         -5.34%
BenchmarkKubeStateMetrics/MakeRequests-16        509309         509375         +0.01%
BenchmarkPodStore-16                             262            232            -11.45%
BenchmarkMetricWrite/value-1-16                  7              7              +0.00%
BenchmarkMetricWrite/value-35.7-16               7              7              +0.00%

benchmark                                        old bytes       new bytes       delta
BenchmarkKubeStateMetrics/GenerateMetrics-16     66918864        63895760        -4.52%
BenchmarkKubeStateMetrics/MakeRequests-16        14257290016     14256884576     -0.00%
BenchmarkPodStore-16                             15888           12816           -19.34%
BenchmarkMetricWrite/value-1-16                  536             536             +0.00%
BenchmarkMetricWrite/value-35.7-16               536             536             +0.00%
```



## Notes
- Benchmarks only test Pods, however the PR changes all `wrap(.*)Func`(s).