@sherifabdlnaby 2022-02-05
I noticed the CI Benchmarks are consistent with mine for **allocs** and **byte/op**, but not for **ns/op**, I ran the benchmarks locally many times and had consistent results though. Thoughts ?