@sherifabdlnaby 2022-02-05
I re-run the benchmarks locally using `GOMAXPROCS=2` too.

Using ` make test-benchmark-compare`

### Compared to Master
<details>
<summary> Logs </summary>

```

### Testing alloc-optimize
?   	k8s.io/kube-state-metrics/v2	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/internal/store
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkPodStore-2   	   99826	     12003 ns/op	   12816 B/op	     232 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/internal/store	4.115s
?   	k8s.io/kube-state-metrics/v2/pkg/allow	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/allowdenylist	2.803s
I0205 20:16:17.412220   98211 builder.go:232] Active resources: certificatesigningrequests,configmaps,cronjobs,daemonsets,deployments,endpoints,horizontalpodautoscalers,ingresses,jobs,leases,limitranges,mutatingwebhookconfigurations,namespaces,networkpolicies,nodes,persistentvolumeclaims,persistentvolumes,poddisruptionbudgets,pods,replicasets,replicationcontrollers,resourcequotas,secrets,services,statefulsets,storageclasses,validatingwebhookconfigurations,volumeattachments
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/app
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkKubeStateMetrics/GenerateMetrics-2         	       1	1001308789 ns/op	63652592 B/op	  602621 allocs/op
BenchmarkKubeStateMetrics/MakeRequests-2            	       1	4241358804 ns/op	1464.03 MB/s	14255555984 B/op	  508637 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/app	8.530s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/builder	1.207s
?   	k8s.io/kube-state-metrics/v2/pkg/builder/types	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/constant	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/customresource	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/metric
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkMetricWrite/value-1-2         	 2952453	       404.7 ns/op	     536 B/op	       7 allocs/op
BenchmarkMetricWrite/value-35.7-2      	 2316003	       518.7 ns/op	     536 B/op	       7 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metric	3.746s
?   	k8s.io/kube-state-metrics/v2/pkg/metric_generator	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metrics_store	0.461s
?   	k8s.io/kube-state-metrics/v2/pkg/metricshandler	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/optin	1.026s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/options	1.029s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/sharding	1.598s
?   	k8s.io/kube-state-metrics/v2/pkg/util/proc	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/version	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/watch	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/e2e	0.437s
?   	k8s.io/kube-state-metrics/v2/tests/e2e/framework	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/lib	0.545s

### Done testing alloc-optimize

### Testing master
Switched to branch 'master'
Your branch is ahead of 'origin/master' by 2 commits.
  (use "git push" to publish your local commits)
?   	k8s.io/kube-state-metrics/v2	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/internal/store
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkPodStore-2   	   85455	     14002 ns/op	   15888 B/op	     262 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/internal/store	1.871s
?   	k8s.io/kube-state-metrics/v2/pkg/allow	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/allowdenylist	0.815s
I0205 20:16:43.182092   98284 builder.go:240] Active resources: certificatesigningrequests,configmaps,cronjobs,daemonsets,deployments,endpoints,horizontalpodautoscalers,ingresses,jobs,leases,limitranges,mutatingwebhookconfigurations,namespaces,networkpolicies,nodes,persistentvolumeclaims,persistentvolumes,poddisruptionbudgets,pods,replicasets,replicationcontrollers,resourcequotas,secrets,services,statefulsets,storageclasses,validatingwebhookconfigurations,volumeattachments
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/app
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkKubeStateMetrics/GenerateMetrics-2         	       1	1003923605 ns/op	66756368 B/op	  636623 allocs/op
BenchmarkKubeStateMetrics/MakeRequests-2            	       1	4389612537 ns/op	1414.58 MB/s	14255488040 B/op	  508684 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/app	8.669s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/builder	0.549s
?   	k8s.io/kube-state-metrics/v2/pkg/builder/types	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/constant	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/customresource	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/metric
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkMetricWrite/value-1-2         	 2975494	       406.7 ns/op	     536 B/op	       7 allocs/op
BenchmarkMetricWrite/value-35.7-2      	 2316817	       520.7 ns/op	     536 B/op	       7 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metric	4.161s
?   	k8s.io/kube-state-metrics/v2/pkg/metric_generator	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metrics_store	1.861s
?   	k8s.io/kube-state-metrics/v2/pkg/metricshandler	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/optin	0.310s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/options	0.446s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/sharding	0.368s
?   	k8s.io/kube-state-metrics/v2/pkg/util/proc	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/version	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/watch	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/e2e	0.332s
?   	k8s.io/kube-state-metrics/v2/tests/e2e/framework	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/lib	1.422s

### Done testing master
Switched to branch 'alloc-optimize'
Your branch is up to date with 'fork/alloc-optimize'.

### Result
old=master new=alloc-optimize
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                       old ns/op      new ns/op      delta
BenchmarkPodStore-2                             14002          12003          -14.28%
BenchmarkKubeStateMetrics/GenerateMetrics-2     1003923605     1001308789     -0.26%
BenchmarkKubeStateMetrics/MakeRequests-2        4389612537     4241358804     -3.38%
BenchmarkMetricWrite/value-1-2                  407            405            -0.49%
BenchmarkMetricWrite/value-35.7-2               521            519            -0.38%

benchmark                                    old MB/s     new MB/s     speedup
BenchmarkKubeStateMetrics/MakeRequests-2     1414.58      1464.03      1.03x

benchmark                                       old allocs     new allocs     delta
BenchmarkPodStore-2                             262            232            -11.45%
BenchmarkKubeStateMetrics/GenerateMetrics-2     636623         602621         -5.34%
BenchmarkKubeStateMetrics/MakeRequests-2        508684         508637         -0.01%
BenchmarkMetricWrite/value-1-2                  7              7              +0.00%
BenchmarkMetricWrite/value-35.7-2               7              7              +0.00%

benchmark                                       old bytes       new bytes       delta
BenchmarkPodStore-2                             15888           12816           -19.34%
BenchmarkKubeStateMetrics/GenerateMetrics-2     66756368        63652592        -4.65%
BenchmarkKubeStateMetrics/MakeRequests-2        14255488040     14255555984     +0.00%
BenchmarkMetricWrite/value-1-2                  536             536             +0.00%
BenchmarkMetricWrite/value-35.7-2               536             536             +0.00%
```
</details>

#### results
```
### Result
old=master new=alloc-optimize
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                       old ns/op      new ns/op      delta
BenchmarkPodStore-2                             14002          12003          -14.28%
BenchmarkKubeStateMetrics/GenerateMetrics-2     1003923605     1001308789     -0.26%
BenchmarkKubeStateMetrics/MakeRequests-2        4389612537     4241358804     -3.38%
BenchmarkMetricWrite/value-1-2                  407            405            -0.49%
BenchmarkMetricWrite/value-35.7-2               521            519            -0.38%

benchmark                                    old MB/s     new MB/s     speedup
BenchmarkKubeStateMetrics/MakeRequests-2     1414.58      1464.03      1.03x

benchmark                                       old allocs     new allocs     delta
BenchmarkPodStore-2                             262            232            -11.45%
BenchmarkKubeStateMetrics/GenerateMetrics-2     636623         602621         -5.34%
BenchmarkKubeStateMetrics/MakeRequests-2        508684         508637         -0.01%
BenchmarkMetricWrite/value-1-2                  7              7              +0.00%
BenchmarkMetricWrite/value-35.7-2               7              7              +0.00%

benchmark                                       old bytes       new bytes       delta
BenchmarkPodStore-2                             15888           12816           -19.34%
BenchmarkKubeStateMetrics/GenerateMetrics-2     66756368        63652592        -4.65%
BenchmarkKubeStateMetrics/MakeRequests-2        14255488040     14255555984     +0.00%
BenchmarkMetricWrite/value-1-2                  536             536             +0.00%
BenchmarkMetricWrite/value-35.7-2               536             536             +0.00%
```



### Compared to Latest Release
<details>
<summary> Logs </summary>

```
./tests/compare_benchmarks.sh release-2.3

### Testing alloc-optimize
?   	k8s.io/kube-state-metrics/v2	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/internal/store
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkPodStore-2   	  100191	     11953 ns/op	   12816 B/op	     232 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/internal/store	1.749s
?   	k8s.io/kube-state-metrics/v2/pkg/allow	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/allowdenylist	0.307s
I0205 20:17:07.754757   98367 builder.go:232] Active resources: certificatesigningrequests,configmaps,cronjobs,daemonsets,deployments,endpoints,horizontalpodautoscalers,ingresses,jobs,leases,limitranges,mutatingwebhookconfigurations,namespaces,networkpolicies,nodes,persistentvolumeclaims,persistentvolumes,poddisruptionbudgets,pods,replicasets,replicationcontrollers,resourcequotas,secrets,services,statefulsets,storageclasses,validatingwebhookconfigurations,volumeattachments
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/app
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkKubeStateMetrics/GenerateMetrics-2         	       1	1000536413 ns/op	63653168 B/op	  602618 allocs/op
BenchmarkKubeStateMetrics/MakeRequests-2            	       1	4355787936 ns/op	1425.57 MB/s	14255397056 B/op	  508674 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/app	9.029s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/builder	0.947s
?   	k8s.io/kube-state-metrics/v2/pkg/builder/types	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/constant	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/customresource	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/metric
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkMetricWrite/value-1-2         	 2972564	       403.1 ns/op	     536 B/op	       7 allocs/op
BenchmarkMetricWrite/value-35.7-2      	 2297282	       521.4 ns/op	     536 B/op	       7 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metric	3.642s
?   	k8s.io/kube-state-metrics/v2/pkg/metric_generator	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metrics_store	0.862s
?   	k8s.io/kube-state-metrics/v2/pkg/metricshandler	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/optin	0.309s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/options	0.345s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/sharding	0.369s
?   	k8s.io/kube-state-metrics/v2/pkg/util/proc	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/version	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/watch	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/e2e	0.333s
?   	k8s.io/kube-state-metrics/v2/tests/e2e/framework	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/lib	0.440s

### Done testing alloc-optimize

### Testing release-2.3
Switched to branch 'release-2.3'
Your branch is up to date with 'origin/release-2.3'.
I0205 20:17:28.223197   98434 builder.go:192] Active resources: certificatesigningrequests,configmaps,cronjobs,daemonsets,deployments,endpoints,horizontalpodautoscalers,ingresses,jobs,leases,limitranges,mutatingwebhookconfigurations,namespaces,networkpolicies,nodes,persistentvolumeclaims,persistentvolumes,poddisruptionbudgets,pods,replicasets,replicationcontrollers,resourcequotas,secrets,services,statefulsets,storageclasses,validatingwebhookconfigurations,volumeattachments
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkKubeStateMetrics/GenerateMetrics-2         	       1	1002657599 ns/op	66786680 B/op	  636612 allocs/op
BenchmarkKubeStateMetrics/MakeRequests-2            	       1	4654870388 ns/op	1333.97 MB/s	14255473344 B/op	  508760 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2	9.108s
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/internal/store
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkPodStore-2   	   83358	     14402 ns/op	   15888 B/op	     262 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/internal/store	1.879s
?   	k8s.io/kube-state-metrics/v2/pkg/allow	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/allowdenylist	0.307s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/builder	0.544s
?   	k8s.io/kube-state-metrics/v2/pkg/builder/types	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/constant	[no test files]
goos: darwin
goarch: amd64
pkg: k8s.io/kube-state-metrics/v2/pkg/metric
cpu: Intel(R) Core(TM) i7-10700 CPU @ 2.90GHz
BenchmarkMetricWrite/value-1-2         	 2943066	       412.7 ns/op	     536 B/op	       7 allocs/op
BenchmarkMetricWrite/value-35.7-2      	 2299832	       524.9 ns/op	     536 B/op	       7 allocs/op
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metric	3.676s
?   	k8s.io/kube-state-metrics/v2/pkg/metric_generator	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/metrics_store	0.358s
?   	k8s.io/kube-state-metrics/v2/pkg/metricshandler	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/optin	0.307s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/options	0.343s
PASS
ok  	k8s.io/kube-state-metrics/v2/pkg/sharding	0.369s
?   	k8s.io/kube-state-metrics/v2/pkg/util/proc	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/version	[no test files]
?   	k8s.io/kube-state-metrics/v2/pkg/watch	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/e2e	1.313s
?   	k8s.io/kube-state-metrics/v2/tests/e2e/framework	[no test files]
PASS
ok  	k8s.io/kube-state-metrics/v2/tests/lib	0.442s

### Done testing release-2.3
Switched to branch 'alloc-optimize'
Your branch is up to date with 'fork/alloc-optimize'.

### Result
old=release-2.3 new=alloc-optimize
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                       old ns/op      new ns/op      delta
BenchmarkKubeStateMetrics/GenerateMetrics-2     1002657599     1000536413     -0.21%
BenchmarkKubeStateMetrics/MakeRequests-2        4654870388     4355787936     -6.43%
BenchmarkPodStore-2                             14402          11953          -17.00%
BenchmarkMetricWrite/value-1-2                  413            403            -2.33%
BenchmarkMetricWrite/value-35.7-2               525            521            -0.67%

benchmark                                    old MB/s     new MB/s     speedup
BenchmarkKubeStateMetrics/MakeRequests-2     1333.97      1425.57      1.07x

benchmark                                       old allocs     new allocs     delta
BenchmarkKubeStateMetrics/GenerateMetrics-2     636612         602618         -5.34%
BenchmarkKubeStateMetrics/MakeRequests-2        508760         508674         -0.02%
BenchmarkPodStore-2                             262            232            -11.45%
BenchmarkMetricWrite/value-1-2                  7              7              +0.00%
BenchmarkMetricWrite/value-35.7-2               7              7              +0.00%

benchmark                                       old bytes       new bytes       delta
BenchmarkKubeStateMetrics/GenerateMetrics-2     66786680        63653168        -4.69%
BenchmarkKubeStateMetrics/MakeRequests-2        14255473344     14255397056     -0.00%
BenchmarkPodStore-2                             15888           12816           -19.34%
BenchmarkMetricWrite/value-1-2                  536             536             +0.00%
BenchmarkMetricWrite/value-35.7-2               536             536             +0.00%
```
</details>

#### results
```
### Result
### Done testing release-2.3
Switched to branch 'alloc-optimize'
Your branch is up to date with 'fork/alloc-optimize'.

### Result
old=release-2.3 new=alloc-optimize
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                       old ns/op      new ns/op      delta
BenchmarkKubeStateMetrics/GenerateMetrics-2     1002657599     1000536413     -0.21%
BenchmarkKubeStateMetrics/MakeRequests-2        4654870388     4355787936     -6.43%
BenchmarkPodStore-2                             14402          11953          -17.00%
BenchmarkMetricWrite/value-1-2                  413            403            -2.33%
BenchmarkMetricWrite/value-35.7-2               525            521            -0.67%

benchmark                                    old MB/s     new MB/s     speedup
BenchmarkKubeStateMetrics/MakeRequests-2     1333.97      1425.57      1.07x

benchmark                                       old allocs     new allocs     delta
BenchmarkKubeStateMetrics/GenerateMetrics-2     636612         602618         -5.34%
BenchmarkKubeStateMetrics/MakeRequests-2        508760         508674         -0.02%
BenchmarkPodStore-2                             262            232            -11.45%
BenchmarkMetricWrite/value-1-2                  7              7              +0.00%
BenchmarkMetricWrite/value-35.7-2               7              7              +0.00%

benchmark                                       old bytes       new bytes       delta
BenchmarkKubeStateMetrics/GenerateMetrics-2     66786680        63653168        -4.69%
BenchmarkKubeStateMetrics/MakeRequests-2        14255473344     14255397056     -0.00%
BenchmarkPodStore-2                             15888           12816           -19.34%
BenchmarkMetricWrite/value-1-2                  536             536             +0.00%
BenchmarkMetricWrite/value-35.7-2               536             536             +0.00%
```